requirements.txt
# flask==3.1.2
# tensorflow==2.20.0
# keras==3.13.1
# numpy==2.4.1
# pillow==12.1.0
# gdown==5.2.0
# gunicorn==21.2.0


tensorflow==2.20.0
keras==3.13.1
numpy
pillow
gradio
